fix(hid): compare keyboard report usages without 16-bit wrap
hid_usages_are_contiguous() compared usages[i] to (uint16_t)(usages[0] + i), which wrapped for usage sequences crossing the 0xffff boundary and could mis-classify a non-contiguous set as contiguous. Compare in uint32_t so the arithmetic does not wrap; valid descriptors parse identically.

+++ src/drivers/class/hid/hid_report_parser.cpp
@@ -90,7 +90,7 @@ static void hid_reset_local_state(HidLocalState *local)
     if (local.usage_count == 0)
         return false;
     for (uint8_t i = 1; i < local.usage_count; i++) {
-        if (local.usages[i] != static_cast<uint16_t>(local.usages[0] + i))
+        if (static_cast<uint32_t>(local.usages[i]) != static_cast<uint32_t>(local.usages[0]) + static_cast<uint32_t>(i))
             return false;
     }
     return true;
